Rename from JavaBridge to JNet
