+++ .github/workflows/codeql-analysis.yml
@@ -83,14 +83,14 @@ jobs:
 
     - name: Maven preparation (step 1)
       if: matrix.language == 'java' 
-      run: dotnet build --no-incremental --configuration Release --framework net5.0 /p:Platform="Any CPU" ./src/net/JavaBridge.sln
+      run: dotnet build --no-incremental --configuration Release --framework net5.0 /p:Platform="Any CPU" ./src/net/JNet.sln
 
     - name: Maven preparation (step 2)
       if: matrix.language == 'java' 
       run: mvn "install:install-file" "-DgroupId=JCOBridge" "-DartifactId=JCOBridge" "-Dversion=2.4.7" "-Dpackaging=jar" "-Dfile=./bin/net5.0/JCOBridge.jar"
 
     - if: matrix.language == 'java' 
-      run: mvn --file ./src/java/javabridge/pom.xml --no-transfer-progress package
+      run: mvn --file ./src/java/jnet/pom.xml --no-transfer-progress package
 
     - name: Perform CodeQL Analysis
       uses: github/codeql-action/analyze@v1
