+++ .github/workflows/pullrequest.yaml
@@ -35,10 +35,10 @@ jobs:
             ${{ runner.os }}-maven-
 
       - name: Pre compile templates
-        run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JavaBridgeTemplates.sln
+        run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JNetTemplates.sln
 
       - name: Pre compile
-        run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JavaBridge\JavaBridge.csproj
+        run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JNet\JNet.csproj
 
       - name: Set up Apache Maven Central
         uses: actions/setup-java@v1
@@ -55,23 +55,23 @@ jobs:
         shell: bash
 
       - name: Create Jars
-        run: mvn --file ./src/java/javabridge/pom.xml --no-transfer-progress package
+        run: mvn --file ./src/java/jnet/pom.xml --no-transfer-progress package
         env:
           MAVEN_USERNAME: ${{ secrets.MAVEN_USERNAME }}
           MAVEN_CENTRAL_TOKEN: ${{ secrets.MAVEN_CENTRAL_TOKEN }}
           MAVEN_GPG_PASSPHRASE: ${{ secrets.MAVEN_GPG_PASSPHRASE }}
 
       # - name: Compile command line
-      #   run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JavaBridgeCLI\JavaBridgeCLI.csproj
+      #   run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JNetCLI\JNetCLI.csproj
       # 
       # - uses: nuget/setup-nuget@v1
       #   with:
       #     nuget-version: '5.x'
       #     
-      # - run: nuget pack src\net\JavaBridgeCLI\JavaBridgeCLI.nuspec -OutputDirectory .\bin
+      # - run: nuget pack src\net\JNetCLI\JNetCLI.nuspec -OutputDirectory .\bin
 
       - name: Recompile to create nuget packages
-        run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JavaBridge.sln
+        run: dotnet build --no-incremental --configuration Release /p:Platform="Any CPU" src\net\JNet.sln
 
       - name: Clear documentation folder 
         run: Remove-Item .\docs\* -Recurse -Force -Exclude _config.yml
