+++ src/net/Documentation/articles/intro.md
@@ -1,5 +1,5 @@
-# JavaBridge: the Java.NET implementation
+# JNet: the Java.NET implementation
 
-JavaBridge is a .NET mirror of Java APIs.
+JNet is a .NET mirror of Java APIs.
 
 This project adheres to the Contributor [Covenant code of conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code. Please report unacceptable behavior to coc_reporting@masesgroup.com.
