+++ src/net/Documentation/articles/usage.md
@@ -1,4 +1,4 @@
-# JavaBridge usage
+# JNet usage
 
-To use JavaBridge classes the developer can write code in .NET using the same classes available in the official Java packages.
+To use JNet classes the developer can write code in .NET using the same classes available in the official Java packages.
 If classes or methods are not available yet it is possible to use the approach synthetized in [What to do if an API was not yet implemented](API_extensibility.md)
