+++ src/net/Documentation/docfx.json
@@ -54,7 +54,7 @@
     "postProcessors": [ "ExtractSearchIndex" ],
     "globalMetadata": {
 	  "_appLogoPath": "images/logo.png",
-      "_appTitle": "Java Bridge website",
+      "_appTitle": "JNet website",
       "_appFooter": "<span>Copyright © 2022 MASES s.r.l..<br>Generated by <strong>DocFX</strong></span>",
       "_gitContribute": {
