+++ src/net/Documentation/index.md
@@ -1,6 +1,6 @@
-# JavaBridge: the Java .NET implementation
+# JNet: the Java .NET implementation
 
-JavaBridge is a .NET mirror of Java APIs.
+JNet is a .NET mirror of Java APIs.
 
 This project adheres to the Contributor [Covenant code of conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code. Please report unacceptable behavior to coc_reporting@masesgroup.com.
 
@@ -9,13 +9,13 @@ This project adheres to the Contributor [Covenant code of conduct](CODE_OF_CONDU
 This project aims to create a library to direct access, from .NET, all the features available in the Java Platform, this is the counterpart of [JCOReflector](https://github.com/masesgroup/JCOReflector)
 
 There are many client libraries written to manage communication with Java. Conversely, this project use directly the Java packages giving more than one benefit:
-* all implemented features are availables at no extra implementation costs, see [JavaBridge usage](src/net/Documentation/articles/usage.md);
+* all implemented features are availables at no extra implementation costs, see [JNet usage](src/net/Documentation/articles/usage.md);
 * avoids any third party communication protocol implementation;
 * access all features made available from Java platform.
 
 ## Runtime engine
 
-JavaBridge uses [JCOBridge](https://www.jcobridge.com), and its [features](https://www.jcobridge.com/features/), to obtain many benefits:
+JNet uses [JCOBridge](https://www.jcobridge.com), and its [features](https://www.jcobridge.com/features/), to obtain many benefits:
 * **Cyber-security**: 
   * [JVM](https://en.wikipedia.org/wiki/Java_virtual_machine) and [CLR, or CoreCLR,](https://en.wikipedia.org/wiki/Common_Language_Runtime) runs in the same process, but are insulated from each other;
   * JCOBridge does not make any code injection into JVM;
@@ -26,7 +26,7 @@ JavaBridge uses [JCOBridge](https://www.jcobridge.com), and its [features](https
   * No need to learn new APIs: we try to expose the same APIs in C# style;
   * No extra validation cycle on protocol and functionality: bug fix, improvements, new features are immediately available;
   * Documentation is shared;
-* **Dynamic code**: it helps to write a Java/Scala/Kotlin/etc seamless language code directly inside a standard .NET application written in C#/VB.NET: look at this [simple example](https://www.jcobridge.com/net-examples/dotnet-examples/) and [JavaBridge APIs extensibility](articles/API_extensibility.md).
+* **Dynamic code**: it helps to write a Java/Scala/Kotlin/etc seamless language code directly inside a standard .NET application written in C#/VB.NET: look at this [simple example](https://www.jcobridge.com/net-examples/dotnet-examples/) and [JNet APIs extensibility](articles/API_extensibility.md).
 
 Have a look at the following resources:
 - [Release notes](https://www.jcobridge.com/release-notes/)
@@ -38,7 +38,7 @@ Have a look at the following resources:
 
 * [Roadmap](articles/roadmap.md)
 * [Actual state](articles/actualstate.md)
-* [JavaBridge usage](articles/usage.md)
-* [JavaBridge APIs extensibility](articles/API_extensibility.md)
+* [JNet usage](articles/usage.md)
+* [JNet APIs extensibility](articles/API_extensibility.md)
 
 ---
