+++ src/net/JNet/JNet.csproj
@@ -1,33 +1,33 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<LangVersion>latest</LangVersion>
-		<AssemblyName>MASES.JavaBridge</AssemblyName>
-		<RootNamespace>MASES.JavaBridge</RootNamespace>
-		<Title>JavaBridge - the Java .NET implementation</Title>
-		<Description>JavaBridge - the Java .NET implementation</Description>
+		<AssemblyName>MASES.JNet</AssemblyName>
+		<RootNamespace>MASES.JNet</RootNamespace>
+		<Title>JNet - the Java .NET implementation</Title>
+		<Description>JNet - the Java .NET implementation</Description>
 		<Copyright>Copyright ©  MASES s.r.l. 2022</Copyright>
 		<Owners>MASES s.r.l.</Owners>
 		<Authors>MASES s.r.l.</Authors>
 		<Company>MASES s.r.l.</Company>
 		<Version>1.0.0.0</Version>
-		<Product>JavaBridge</Product>
+		<Product>JNet</Product>
 		<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
 		<TargetFrameworks>net461;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
 		<OutputPath>..\..\..\bin\</OutputPath>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-		<PackageProjectUrl>https://github.com/masesgroup/JavaBridge/</PackageProjectUrl>
-		<RepositoryUrl>https://github.com/masesgroup/JavaBridge</RepositoryUrl>
-		<PackageReleaseNotes>https://github.com/masesgroup/JavaBridge/releases</PackageReleaseNotes>
+		<PackageProjectUrl>https://github.com/masesgroup/JNet/</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/masesgroup/JNet</RepositoryUrl>
+		<PackageReleaseNotes>https://github.com/masesgroup/JNet/releases</PackageReleaseNotes>
 		<PackageLicenseFile>LICENSE</PackageLicenseFile>
 		<PackageIcon>JCOB128x128.png</PackageIcon>
-		<PackageTags>javabridge dotnet clr netcore net5 net6 java</PackageTags>
-		<PackageId>MASES.JavaBridge</PackageId>
+		<PackageTags>jnet dotnet clr netcore net5 net6 java bridge</PackageTags>
+		<PackageId>MASES.JNet</PackageId>
 		<IncludeSymbols>true</IncludeSymbols>
 		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
 		<EmbedUntrackedSources>true</EmbedUntrackedSources>
 		<SignAssembly>true</SignAssembly>
-		<AssemblyOriginatorKeyFile>..\Common\JavaBridge.snk</AssemblyOriginatorKeyFile>
+		<AssemblyOriginatorKeyFile>..\Common\JNet.snk</AssemblyOriginatorKeyFile>
 		<PackageReadmeFile>usage.md</PackageReadmeFile>
 	</PropertyGroup>
   <!-- Fix start https://github.com/dotnet/sourcelink/issues/572 -->
@@ -47,7 +47,7 @@
   <ItemGroup>
     <None Include="..\..\..\LICENSE" Pack="true" PackagePath="" />
     <None Include="..\Common\JCOB128x128.png" Pack="true" PackagePath="" />
-    <None Include="..\Common\JavaBridge.snk" Link="JavaBridge.snk" />
+    <None Include="..\Common\JNet.snk" Link="JNet.snk" />
     <None Include="..\Documentation\articles\usage.md" Pack="true" PackagePath="\" />
   </ItemGroup>
   <ItemGroup>
