+++ src/net/JNet/Java/Math/BigDecimal.cs
