+++ src/net/JNet/Java/Time/Duration.cs
