+++ src/net/JNet/Java/Util/AbstractMap.cs
