+++ src/net/JNet/Java/Util/Collection.cs
