+++ src/net/JNet/Java/Util/Collections.cs
