+++ src/net/JNet/Java/Util/Concurrent/Future.cs
