+++ src/net/JNet/Java/Util/Dictionary.cs
