+++ src/net/JNet/Java/Util/HashMap.cs
