+++ src/net/JNet/Java/Util/Iterator.cs
