+++ src/net/JNet/Java/Util/Map.cs
