+++ src/net/JNet/Java/Util/Optional.cs
