+++ src/net/JNet/Java/Util/OptionalDouble.cs
