+++ src/net/JNet/Java/Util/OptionalInt.cs
