+++ src/net/JNet/Java/Util/OptionalLong.cs
