+++ src/net/JNet/Java/Util/Properties.cs
