+++ src/net/JNet/Java/Util/RegEx/Duration.cs
