+++ src/net/JNet/Java/Util/Set.cs
