+++ src/net/JNet/Java/Util/UUID.cs
