--- src/net/JavaBridgeTemplates.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32014.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JavaBridge", "JavaBridge\JavaBridge.csproj", "{E7C8F876-71A0-4EE0-A4E9-58939954A1CC}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "javabridgeApp", "templates\templates\javabridgeApp\javabridgeApp.csproj", "{C47B3A03-793E-4102-A872-4A39C6E03575}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E7C8F876-71A0-4EE0-A4E9-58939954A1CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E7C8F876-71A0-4EE0-A4E9-58939954A1CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E7C8F876-71A0-4EE0-A4E9-58939954A1CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E7C8F876-71A0-4EE0-A4E9-58939954A1CC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C47B3A03-793E-4102-A872-4A39C6E03575}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C47B3A03-793E-4102-A872-4A39C6E03575}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C47B3A03-793E-4102-A872-4A39C6E03575}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C47B3A03-793E-4102-A872-4A39C6E03575}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {337100A8-DBCD-4838-B9C7-74EC0DE78755}
-	EndGlobalSection
-EndGlobal
