+++ src/net/templates/templatepack.csproj
@@ -2,29 +2,29 @@
   <PropertyGroup>
     <PackageType>Template</PackageType>
     <PackageVersion>1.0.0.0</PackageVersion>
-    <PackageId>MASES.JavaBridge.Templates</PackageId>
-    <Title>JavaBridge Templates - Templates to use the JavaBridge</Title>
+    <PackageId>MASES.JNet.Templates</PackageId>
+    <Title>JNet Templates - Templates to use the JNet</Title>
     <Authors>MASES s.r.l.</Authors>
     <Owners>MASES s.r.l.</Owners>
     <Company>MASES s.r.l.</Company>
     <Copyright>Copyright ©  MASES s.r.l. 2022</Copyright>
-    <Summary>Ready made templates to create applications based on JavaBridge</Summary>
-    <Description>Ready made templates to create applications based on JavaBridge. The templates are ready made starting points, all information and APIs are available in the official website https://masesgroup.github.io/JavaBridge</Description>
+    <Summary>Ready made templates to create applications based on JNet</Summary>
+    <Description>Ready made templates to create applications based on JNet. The templates are ready made starting points, all information and APIs are available in the official website https://masesgroup.github.io/JNet</Description>
     <TargetFrameworks>net6.0;net5.0;netcoreapp3.1;net461</TargetFrameworks>
     <OutputPath>..\..\..\bin\</OutputPath>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <PackageProjectUrl>https://github.com/masesgroup/JavaBridge/</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/masesgroup/JavaBridge</RepositoryUrl>
-    <PackageReleaseNotes>https://github.com/masesgroup/JavaBridge/releases</PackageReleaseNotes>
+    <PackageProjectUrl>https://github.com/masesgroup/JNet/</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/masesgroup/JNet</RepositoryUrl>
+    <PackageReleaseNotes>https://github.com/masesgroup/JNet/releases</PackageReleaseNotes>
 	<PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageIcon>JCOB128x128.png</PackageIcon>
     <PackageTags>javabridge java dotnet clr netcore net5 net6 template</PackageTags>
     <LangVersion>8.0</LangVersion>
-    <Product>JavaBridge Templates</Product>
+    <Product>JNet Templates</Product>
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
     <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>..\Common\JavaBridge.snk</AssemblyOriginatorKeyFile>
+    <AssemblyOriginatorKeyFile>..\Common\JNet.snk</AssemblyOriginatorKeyFile>
     <IncludeContentInPack>true</IncludeContentInPack>
     <IncludeBuildOutput>false</IncludeBuildOutput>
     <ContentTargetFolders>content</ContentTargetFolders>
@@ -44,7 +44,7 @@
   <ItemGroup>
     <None Include="..\..\..\LICENSE" Pack="true" PackagePath="" />
     <None Include="..\Common\JCOB128x128.png" Pack="true" PackagePath="" />
-    <None Include="..\Common\JavaBridge.snk" Link="JavaBridge.snk" />
+    <None Include="..\Common\JNet.snk" Link="JNet.snk" />
     <None Include="..\..\..\README.md" Pack="true" PackagePath="\" />
     <Content Include="templates\**\*" Exclude="templates\**\bin\**;templates\**\obj\**;templates\**\.vscode\**" />
     <Compile Remove="**\*" />
