+++ src/net/templates/templates/jnetApp/.template.config/template.json
@@ -2,9 +2,9 @@
   "$schema": "http://json.schemastore.org/template",
   "author": "MASES s.r.l.",
   "classifications": [ "Common", "Console", "C#8" ],
-  "identity": "MASES.JavaBridgeTemplate.JavaBridgeApp",
-  "name": "Console templates: Java Bridge project",
-  "shortName": "javabridgeApp",
+  "identity": "MASES.JNetTemplate.JNetApp",
+  "name": "Console templates: JNet project",
+  "shortName": "jnetApp",
   "tags": {
     "language": "C#",
     "type": "project"
