+++ src/net/templates/templates/jnetApp/jnetApp.csproj
@@ -6,12 +6,12 @@
   <ItemGroup Condition="'$(TargetFramework)' == 'net461'">
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
-  <ItemGroup Condition="Exists('..\..\..\JavaBridge\JavaBridge.csproj')">
+  <ItemGroup Condition="Exists('..\..\..\JNet\JNet.csproj')">
     <!--Within GitHub repo: used for test purpose-->
-    <ProjectReference Include="..\..\..\JavaBridge\JavaBridge.csproj" />
+    <ProjectReference Include="..\..\..\JNet\JNet.csproj" />
   </ItemGroup>
-  <ItemGroup Condition="!Exists('..\..\..\JavaBridge\JavaBridge.csproj')">
+  <ItemGroup Condition="!Exists('..\..\..\JNet\JNet.csproj')">
     <!--Outside GitHub repo-->
-    <PackageReference Include="MASES.JavaBridge" Version="1.0.0" IncludeAssets="All" PrivateAssets="None" />
+    <PackageReference Include="MASES.JNet" Version="1.0.0" IncludeAssets="All" PrivateAssets="None" />
   </ItemGroup>
 </Project>
