+++ tests/JNetTest.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JNet", "..\src\net\JNet\JNet.csproj", "{067E5A2E-168E-4CD5-855F-672A0FE86DA1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JNetTest", "JNetTest\JNetTest.csproj", "{F3A53651-9728-427B-A429-7B3F1B6C8DA0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{067E5A2E-168E-4CD5-855F-672A0FE86DA1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{067E5A2E-168E-4CD5-855F-672A0FE86DA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{067E5A2E-168E-4CD5-855F-672A0FE86DA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{067E5A2E-168E-4CD5-855F-672A0FE86DA1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F3A53651-9728-427B-A429-7B3F1B6C8DA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3A53651-9728-427B-A429-7B3F1B6C8DA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3A53651-9728-427B-A429-7B3F1B6C8DA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3A53651-9728-427B-A429-7B3F1B6C8DA0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0A7C16DC-1BAA-44BC-AA1C-D40B7B61878E}
+	EndGlobalSection
+EndGlobal
