+++ tests/JNetTest/Program.cs
@@ -20,13 +20,13 @@ using Java.Util;
 using MASES.JavaBridge;
 using System;
 
-namespace MASES.JavaBridgeTemplate.JavaBridgeApp
+namespace MASES.JNetTest
 {
     class Program
     {
         static void Main(string[] args)
         {
-            var appArgs = JavaBridgeCore.GlobalInstance.ApplicationArgs;
+            var appArgs = JNetCore.GlobalInstance.ApplicationArgs;
             if (args.Length != 0)
             {
                 var set = Collections.Singleton(appArgs[0]);
