--- tests/JavaBridgeTest.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32014.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JavaBridge", "..\src\net\JavaBridge\JavaBridge.csproj", "{07A62D28-DDB1-489A-844F-49B5B0C30CAE}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JavaBridgeTest", "JavaBridgeTest\JavaBridgeTest.csproj", "{DCB70ACA-DDF3-4821-93D6-9063618C421C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{07A62D28-DDB1-489A-844F-49B5B0C30CAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{07A62D28-DDB1-489A-844F-49B5B0C30CAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{07A62D28-DDB1-489A-844F-49B5B0C30CAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{07A62D28-DDB1-489A-844F-49B5B0C30CAE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DCB70ACA-DDF3-4821-93D6-9063618C421C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DCB70ACA-DDF3-4821-93D6-9063618C421C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DCB70ACA-DDF3-4821-93D6-9063618C421C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DCB70ACA-DDF3-4821-93D6-9063618C421C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0A7C16DC-1BAA-44BC-AA1C-D40B7B61878E}
-	EndGlobalSection
-EndGlobal
